refactor: rename to .ts & refactor indents in sources file
